==> baseball_retrosheet/0. Unzip.ipynb <==
import os
import zipfile
events_dir = '/home/idies/workspace/Temporary/raddick/baseball_scratch/events/'
games_dir = '/home/idies/workspace/Temporary/raddick/baseball_scratch/games/'
os.chdir(events_dir)
print('ok')
os.getcwd()

filelist = os.listdir()
filelist = sorted(filelist)
#filelist
for thisfile in filelist:
    print('Extracting {0:}...'.format(thisfile))
    thezip = zipfile.ZipFile(thisfile)
    thezip.extractall()
print('Done!')


os.chdir(games_dir)
filelist = os.listdir()
filelist = sorted(filelist)
#filelist
for thisfile in filelist:
    print('Extracting {0:}...'.format(thisfile))
    thezip = zipfile.ZipFile(thisfile)
    thezip.extractall()
print('Done!')
