==> baseball_retrosheet/1. Get rosters.ipynb <==
import os
import numpy as np
import pandas
os.getcwd()

os.chdir('/home/idies/workspace/Storage/raddick/baseball/events/')
filelist = []
for thisfile in os.listdir():
    filename = str(thisfile)
#    print(filename)
    if ('.ROS' in filename):
        filelist.append('events/'+filename)

os.chdir('/home/idies/workspace/Storage/raddick/baseball/')
#print(filelist)
print('files identified')

#filelist = ['events/MLN1954.ROS']
colnames = ['playerID', 'nameLast', 'nameFirst', 'batting_hand', 'throwing_hand', 'team', 'position', 'year']

roster_df = pandas.DataFrame(columns=colnames)

for i in range(0,len(filelist)):
    if ((np.mod(i,500) == 0) | (i == len(filelist))):
        print('Parsing roster {0:,.0f} of {1:,.0f}...'.format(i, len(filelist)))
    filename = filelist[i]
    thisyear = str(filename).split('/')[1].split('.')[0][3:7]
    this_roster_df = pandas.read_csv(filename, names=colnames, header=None)
    #this_roster_df = this_roster_df.set_index('playerID')
    this_roster_df['year'] = thisyear
    roster_df = pandas.concat((roster_df, this_roster_df)) 

roster_df['year'] = pandas.to_numeric(roster_df['year'], downcast='integer', errors='coerce')

new_column_order = ['playerID', 'nameLast', 'nameFirst', 'batting_hand', 'throwing_hand', 'position', 'team', 'year']
roster_df = roster_df[new_column_order]

roster_df = roster_df.sort_values(['playerID','year', 'team'])
roster_df = roster_df.reset_index(drop=True)
roster_df.to_csv('fullroster.csv', encoding='utf-8')
print('Done')
