==> baseball_retrosheet/2. Load game data.ipynb <==
#import zipfile
import numpy as np
import pandas
import os
pandas.options.display.max_colwidth = -1
os.chdir('/home/idies/workspace/Storage/raddick/baseball/games/')
print('ok')

years = range(1871,2017+1)
filenames = []
print('Finding filenames...')
for thisyear in years:
    filenames.append('gl'+str(thisyear)+'.txt')

#filenames
gamedata = pandas.DataFrame()

print('Reading file contents...')
list_of_file_contents = []
for thisname in filenames:
#    print('Reading filename: ',thisname)
    df = pandas.read_csv(thisname,header=None)
    list_of_file_contents.append(df)
gamedata = pandas.concat(list_of_file_contents)

# Columns given at http://www.retrosheet.org/gamelogs/glfields.txt
colnames = ['date','gameno','dayofweek']
colnames += ['visitor','visitor_league','visitor_gameno']
colnames += ['home','home_league','home_gameno']
colnames += ['visitor_score','home_score']
colnames += ['nOuts','day_night','completed','forfeit','protest','parkID','attendance','length']
colnames += ['visitor_linescore','home_linescore']
colnames += ['visitor_ab','visitor_h','visitor_2b','visitor_3b','visitor_hr']
colnames += ['visitor_rbi','visitor_sac','visitor_sacfly','visitor_hbp']
colnames += ['visitor_bb','visitor_ibb','visitor_k','visitor_sb','visitor_cs']
colnames += ['visitor_gidp','visitor_ci','visitor_lob','visitor_nPitchers']
colnames += ['visitor_er_ind','visitor_er_team','visitor_wp','visitor_balk']
colnames += ['visitor_putout','visitor_assist','visitor_e','visitor_pb']
colnames += ['visitor_dp','visitor_tp','home_ab','home_h','home_2b','home_3b']
colnames += ['home_hr','home_rbi','home_sac','home_sacfly','home_hbp','home_bb']
colnames += ['home_ibb','home_k','home_sb','home_cs','home_gidp','home_ci']
colnames += ['home_lob','home_nPitchers','home_er_ind','home_er_team','home_wp']
colnames += ['home_balk','home_putout','home_assist','home_e','home_pb']
colnames += ['home_dp','home_tp']
colnames += ['ump_home_id','ump_home_name','ump_1b_id']
colnames += ['ump_1b_name','ump_2b_id','ump_2b_name','ump_3b_id','ump_3b_name']
colnames += ['ump_lf_id','ump_lf_name','ump_rf_id','ump_rf_name']
colnames += ['visitor_manager_id','visitor_manager_name','home_manager_id','home_manager_name']
colnames += ['winning_pitcher_id','winning_pitcher_name','losing_pitcher_id']
colnames += ['losing_pitcher_name','saving_pitcher_id','saving_pitcher_name']
colnames += ['gw_rbi_batter_id','gw_rbi_batter_name','visitor_starting_pitcher_id']
colnames += ['visitor_starting_pitcher_name','home_starting_pitcher_id']
colnames += ['home_starting_pitcher_name']
colnames += ['visitor_batter1_id','visitor_batter1_name','visitor_batter1_pos']
colnames += ['visitor_batter2_id','visitor_batter2_name','visitor_batter2_pos']
colnames += ['visitor_batter3_id','visitor_batter3_name','visitor_batter3_pos']
colnames += ['visitor_batter4_id','visitor_batter4_name','visitor_batter4_pos']
colnames += ['visitor_batter5_id','visitor_batter5_name','visitor_batter5_pos']
colnames += ['visitor_batter6_id','visitor_batter6_name','visitor_batter6_pos']
colnames += ['visitor_batter7_id','visitor_batter7_name','visitor_batter7_pos']
colnames += ['visitor_batter8_id','visitor_batter8_name','visitor_batter8_pos']
colnames += ['visitor_batter9_id','visitor_batter9_name','visitor_batter9_pos']
colnames += ['home_batter1_id','home_batter1_name','home_batter1_pos']
colnames += ['home_batter2_id','home_batter2_name','home_batter2_pos']
colnames += ['home_batter3_id','home_batter3_name','home_batter3_pos']
colnames += ['home_batter4_id','home_batter4_name','home_batter4_pos']
colnames += ['home_batter5_id','home_batter5_name','home_batter5_pos']
colnames += ['home_batter6_id','home_batter6_name','home_batter6_pos']
colnames += ['home_batter7_id','home_batter7_name','home_batter7_pos']
colnames += ['home_batter8_id','home_batter8_name','home_batter8_pos']
colnames += ['home_batter9_id','home_batter9_name','home_batter9_pos']
colnames += ['misc','acquisition']

gamedata.columns = colnames

print('Calculating gameID...')
gamedata = gamedata.assign(gameID = gamedata.apply(lambda row: row['home']+str(row['date'])+str(row['gameno']), axis=1))
gamedata = gamedata.set_index('gameID')

print('Fixing data types...')
gamedata['date'] = pandas.to_datetime(gamedata['date'], format='%Y%m%d', errors='coerce')

gamedata['visitor_linescore'] = gamedata['visitor_linescore'].astype('str')
gamedata['home_linescore'] = gamedata['home_linescore'].astype('str')
gamedata.loc[gamedata['visitor_linescore'] == 'nan', 'visitor_linescore'] = 'n/a'
gamedata.loc[gamedata['home_linescore'] == 'nan', 'home_linescore'] = 'n/a'

print('Writing {:,.0f} games to CSV...'.format(len(gamedata)))

gamedata.to_csv('/home/idies/workspace/Temporary/raddick/baseball_scratch/gamedata.csv')
print('Done!')
